requirements.txt
africastalking==1.1.5
aioredis==1.3.0
asgiref==3.2.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.7.0
beautifulsoup4==4.8.1
certifi==2019.3.9
cffi==1.13.0
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2
django-bootstrap4==1.1.0
django-cors-headers==2.5.2
django-extensions==2.2.1
django-filter==2.1.0
django-richtextfield==1.5.0
django-tenant-schemas==1.9.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1
msgpack==0.6.2
Pillow==6.0.0
psycopg2==2.8.2
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.2
python-decouple==3.1
pytz==2018.9
requests==2.22.0
schedule==0.6.0
schema==0.7.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.2
whitenoise==4.1.4
zope.interface==4.6.0
